+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on Keep a Changelog and the project follows
 Conventional Commits.
 
+## [Unreleased]
+
+### Added
+- Auto completion for file paths to open files in the command line prompt.
+
 ## [0.1.0] - 2025-12-30
 
 ### Added